schedule an rating and desc

+++ Model/Teacher.cs
@@ -8,13 +8,13 @@ namespace Model
 {
     public class Teacher : User
     {
-        private int rating;
-        private int price;
+        private double rating;
+        private double price;
         private int amountOfJobs;
         private string description;
 
-        public int Rating { get => rating; set => rating = value; }
-        public int Price { get => price; set => price = value; }
+        public double Rating { get => rating; set => rating = value; }
+        public double Price { get => price; set => price = value; }
         public int AmountOfJobs { get => amountOfJobs; set => amountOfJobs = value; }
         public string Description { get => description; set => description = value; }
     }
